Move StepSlider to design package (connect will also use it)
