+++ packages/design/src/StepSlider/StepSlider.test.tsx
