+++ packages/design/src/index.ts
@@ -38,6 +38,7 @@ import Link from './Link';
 import Image from './Image';
 import Text from './Text';
 import SideNav, { SideNavItem } from './SideNav';
+import { StepSlider } from './StepSlider';
 import TopNav, { TopNavItem } from './TopNav';
 import Flex from './Flex';
 import Menu, { MenuItem, MenuItemIcon } from './Menu';
@@ -73,6 +74,7 @@ export {
   Image,
   SideNav,
   SideNavItem,
+  StepSlider,
   Text,
   TopNav,
   TopNavItem,
