+++ packages/teleport/src/Welcome/NewCredentials/NewCredentials.tsx
@@ -18,11 +18,7 @@ import React, { useState } from 'react';
 import { Card } from 'design';
 import { PrimaryAuthType } from 'shared/services';
 import RecoveryCodes from 'teleport/components/RecoveryCodes';
-import {
-  StepSlider,
-  NewFlow,
-  StepComponentProps,
-} from 'teleport/components/StepSlider';
+import { StepSlider, NewFlow, StepComponentProps } from 'design/StepSlider';
 import useToken, { State } from '../useToken';
 import { Expired } from './Expired';
 import { RegisterSuccess } from './Success';
